refactor: Move tests to DiagramHelpersTest from LineTest
This matches the location of the implementation functions. Those functions were moved some time back and the associated tests were never moved along with them.
